# Performance of appending option to select