--- a/benchmark/html/index.js
+++ b/benchmark/html/index.js
@@ -1,5 +1,6 @@
 "use strict";
 
 module.exports = {
-  parsing: require("./parsing")
+  parsing: require("./parsing"),
+  select: require("./select")
 };