--- a/benchmark/html/select.js
+++ b/benchmark/html/select.js
@@ -1,0 +1,9 @@
+"use strict";
+const suite = require("../document-suite");
+
+exports.text = suite(document => {
+  document.body.innerHTML = `
+  <select>
+    ${"<option value=\"volvo\">Volvo</option>".repeat(5000)}
+  </select>`;
+});