--- a/lib/jsdom/living/nodes/HTMLSelectElement-impl.js
+++ b/lib/jsdom/living/nodes/HTMLSelectElement-impl.js
@@ -15,6 +15,19 @@ const { domSymbolTree } = require("../helpers/internal-constants");
 const { getLabelsForLabelable, formOwner, isDisabled } = require("../helpers/form-controls");
 const { parseNonNegativeInteger } = require("../helpers/strings");
 
+function toOptionArray(child) {
+  const array = [];
+  if (child._localName === "option" && !child.hasAttributeNS(null, "disabled")) {
+    array.push(child);
+  } else if (child._localName === "optgroup") {
+    for (const childOfGroup of domSymbolTree.childrenIterator(child)) {
+      if (childOfGroup._localName === "option" && !childOfGroup.hasAttributeNS(null, "disabled")) {
+        array.push(childOfGroup);
+      }
+    }
+  }
+  return array;
+}
 class HTMLSelectElementImpl extends HTMLElementImpl {
   constructor(globalObject, args, privateData) {
     super(globalObject, args, privateData);
@@ -42,6 +55,10 @@ class HTMLSelectElementImpl extends HTMLElementImpl {
     this._customValidityErrorMessage = "";
 
     this._labels = null;
+
+    // Essentially a cache for the current selected element for use while building a select by appending options
+    // See #_selectednessSettingAlgorithmWithKnownNewAddition
+    this._lastSelectedOptionForFastPath = null;
   }
 
   _formReset() {
@@ -53,6 +70,10 @@ class HTMLSelectElementImpl extends HTMLElementImpl {
   }
 
   _askedForAReset() {
+    this._selectednessSettingAlgorithm();
+  }
+
+  _selectednessSettingAlgorithm() {
     if (this.hasAttributeNS(null, "multiple")) {
       return;
     }
@@ -74,6 +95,7 @@ class HTMLSelectElementImpl extends HTMLElementImpl {
         if (!disabled) {
           // (do not set dirty)
           option._selectedness = true;
+          this._lastSelectedOptionForFastPath = option;
           break;
         }
       }
@@ -82,12 +104,46 @@ class HTMLSelectElementImpl extends HTMLElementImpl {
       selected.forEach((option, index) => {
         option._selectedness = index === selected.length - 1;
       });
+      this._lastSelectedOptionForFastPath = selected[selected.length - 1];
+    }
+  }
+
+  _selectednessSettingAlgorithmWithKnownNewAddition(child) {
+    if (this.hasAttributeNS(null, "multiple")) {
+      return;
+    }
+    if (child.hasAttributeNS(null, "disabled")) {
+      return;
+    }
+    const optArr = toOptionArray(child);
+    if (optArr.length === 0) {
+      return;
+    }
+    const selectedOptions = optArr.filter(opt => opt._selectedness);
+    // If any selected is last selected
+    if (selectedOptions.length > 0) {
+      // pushed at least one element with selected=true, reset
+      this._askedForAReset();
+      return;
+    }
+
+    // No selection so first option is selected
+    if ((!this._lastSelectedOptionForFastPath || !this._lastSelectedOptionForFastPath._selectedness) &&
+        this._displaySize === 1) {
+      this._lastSelectedOptionForFastPath = optArr[0];
+      this._lastSelectedOptionForFastPath._selectedness = true;
     }
   }
 
   _descendantAdded(parent, child) {
     if (child.nodeType === NODE_TYPE.ELEMENT_NODE) {
-      this._askedForAReset();
+      if (parent === this) {
+        this._selectednessSettingAlgorithmWithKnownNewAddition(child);
+      } else if (parent._localName === "optgroup") {
+        this._selectednessSettingAlgorithmWithKnownNewAddition(parent);
+      } else {
+        this._askedForAReset();
+      }
     }
 
     super._descendantAdded(parent, child);
@@ -95,6 +151,9 @@ class HTMLSelectElementImpl extends HTMLElementImpl {
 
   _descendantRemoved(parent, child) {
     if (child.nodeType === NODE_TYPE.ELEMENT_NODE) {
+      if (this._lastSelectedOptionForFastPath === child) {
+        this._lastSelectedOptionForFastPath = null;
+      }
       this._askedForAReset();
     }