--- a/lib/jsdom/living/nodes/HTMLSelectElement-impl.js
+++ b/lib/jsdom/living/nodes/HTMLSelectElement-impl.js
@@ -15,19 +15,6 @@ const { domSymbolTree } = require("../helpers/internal-constants");
 const { getLabelsForLabelable, formOwner, isDisabled } = require("../helpers/form-controls");
 const { parseNonNegativeInteger } = require("../helpers/strings");
 
-function toOptionArray(child) {
-  const array = [];
-  if (child._localName === "option" && !child.hasAttributeNS(null, "disabled")) {
-    array.push(child);
-  } else if (child._localName === "optgroup") {
-    for (const childOfGroup of domSymbolTree.childrenIterator(child)) {
-      if (childOfGroup._localName === "option" && !childOfGroup.hasAttributeNS(null, "disabled")) {
-        array.push(childOfGroup);
-      }
-    }
-  }
-  return array;
-}
 class HTMLSelectElementImpl extends HTMLElementImpl {
   constructor(globalObject, args, privateData) {
     super(globalObject, args, privateData);
@@ -56,8 +43,8 @@ class HTMLSelectElementImpl extends HTMLElementImpl {
 
     this._labels = null;
 
-    // Essentially a cache for the current selected element for use while building a select by appending options
-    // See #_selectednessSettingAlgorithmWithKnownNewAddition
+    // Essentially a cache for the current selected element for use while building a select by appending options.
+    // See _selectednessSettingAlgorithmWithKnownNewAddition().
     this._lastSelectedOptionForFastPath = null;
   }
 
@@ -336,6 +323,20 @@ class HTMLSelectElementImpl extends HTMLElementImpl {
 }
 
 mixin(HTMLSelectElementImpl.prototype, DefaultConstraintValidationImpl.prototype);
+
+function toOptionArray(child) {
+  const array = [];
+  if (child._localName === "option" && !child.hasAttributeNS(null, "disabled")) {
+    array.push(child);
+  } else if (child._localName === "optgroup") {
+    for (const childOfGroup of domSymbolTree.childrenIterator(child)) {
+      if (childOfGroup._localName === "option" && !childOfGroup.hasAttributeNS(null, "disabled")) {
+        array.push(childOfGroup);
+      }
+    }
+  }
+  return array;
+}
 
 module.exports = {
   implementation: HTMLSelectElementImpl