@eps1lon 2022-12-29
This broke `HTMLSelectElement.selectedOptions` if `HTMLOptionElement.selected` is used: https://github.com/jsdom/jsdom/issues/3476

Revert in progress in https://github.com/jsdom/jsdom/pull/3475

Would be nice to reland with the new test passing though.